Continued work on init page
Now allows for uploading of .json and .csv files from user. To accommodate this change, init_options requirements were changed from requiring just a JSON string as a dataset to requiring an Object with the string as its single attribute, "jsonString". Modified the json_string page to fit this requirement.