Beginning of a FilteredColumn decorator class
Setup a simple class that holds a data column and necessary filter information from a graph view. Currently this class can: return the data from the column filtered through the filters, return the original column added in the constructor, reset the filter requirements of the FilteredColumn, and add new filter requirements to  FilteredColumn.
Changed the init options for the index page to start with a Distribution Dot Plot for more obvious filtering for debugging. 
Added the file requirements for the decorator file into the SeeIt.js and application.js.
Some additions to graph_view for testing, not necessary for functionality.