Filter Decorator additions
Added functions in the decorator so that every function in a normal DataColumn has a counterpart in the FilteredColumn. For some of the functions there are calls to make the relevant call on the original DataColumn or to perform the same action but on the FilteredColumn(toJson, uniqueData, length)
A couple more changes to GraphView to ensure complete compatibility with the filteredColumn, specifically for the tooltip that appears displaying filter statistics when clicking on the buttons for each assigned dataColumn in the footer.