Small fix and cleanup
Modified a previous change to RemoveData in graph_view where it was looking down into the DataColumn of the FilteredColumns and instead had it look directly for the FilteredColumn when searching to remove.
Cleaned up the commented out legacy lines that are no longer used because of the filteredColumn dataset in graph view.