Couple Fixes
StackedAreaChart and BarChart now both getColor correctly for each DataColumn assigned to them.
Fixed a bug where data footers would be removed whenever new data was added to the footer if that data role was marked "multiple = false", regardless of if the removed data was in the role that had new data being assigned to it.