Added some comments, changed some names
Changed some of the comments in graph.js to reflect that FilteredColumns would be what is actually passed into the graph for data.
Changed the function names for mimiced functions in the filtered_column to be filtered{function_name_here} and {function_name_here} for the filtered version and original, respectively, instead of {function_name_here} and original{function_name_here} so that the same calls would have consistent behavior throughout the code (though the mimiced functions are only called in parts of the code that were not changed to now deal with FilteredColumns, so this probably won't matter).
Removed the filtered_column_factory.js file as it is no longer used, and renamed the filter_decorator.js file to filtered_column.js, as well as changing that filename in application.js and SeeIt.js