The app should now wait until all data is loaded before initing graphs
DataCollection now keeps track of all datasets that are a part of the initialization data through an array, pendingDatasets, which stores the urls of each remote endpoint.
If this array is empty after its attempted population, a variable is set in the ApplicationController, graphGoAhead, which tells it to continue as it was before as there are no issues waiting for download callbacks, all the init data is given to the app up front.
In initDatasets, the coerceDataset callback now accepts an additional argument, cbUrl, which is the url of the dataset which is returning from its download .  If this argument is not null, the callback will remove the url from pendingDatasets and check if the array is now empty.
If pendingDatasets is empty, it triggers the event 'datasets:loaded', which the ApplicationController listens for. When the App hears it, it calls the initGraphs function which creates each graph that was in the initialization params. 
As the DataCollectionView initializes with the DataCollection model, I changed it so that the view would not be initialized unless all data had been loaded to the model. The registerListeners() function in AppController relies on the DataCollectionView being initialized, so because of these dependencies, both of those calls(along with the trigger('ready')) were moved within the 'datasets:loaded' event or the "if @graphGoAhead" block at the end of the constructor.