Two Equal Division added to DDP
Added a select option to the DistributionDotPlot for dividers on the graph a la the original SeeIt. Currently only options are "None" and "Two Equal". Planned additions to this option are also "Four Equal" and "Choose Your Own".
Two Equal option right now will draw lines at the minimum valued point on the graph, the maximum value point, and at a point somewhere in the graph that divides the graph into two sections with an equal distribution of points on each side of it. The amount of points in each section is displayed near the top of the graph in the middle of each section. If there is an odd number of data points in the graph, the smaller division is on the left.
Fixed a bug in the graph options where the value attribute of <option> tags in graphOptions with the type "select" was given unexpected values if the string describing the value contained spaces.
Also made a small fix in regards to filteredColumns that was a result of a naming convention change of functions that have a filtered version and an unfiltered version.