Added options for Four Equal divs and ones of fixed size
With the addition of multiple options drawing the divs, I made each option call the drawDivs function, the different option values result in different data setup, but the drawing of divs is handled the same d3 code regardless.
Four Equal is basically the same as Two Equal, has the same behavior that if the divisions are not of equal size, aka the set is not of a size divisible by 4, then the smaller divisions will be on the left.
The fixed size option is given a number, n, then divisions are made by going through the data and making a dividing line after every nth point. If the set is not cleanly divisible by n, the last division will just be filled with the remainder points.
Because the Fixed Size Dividers option requires a text input, it is in its own GraphOption. This should ideally be like an option in the dividers select, i.e. it cannot be enabled while the other dividers or box plot etc. are on, but that will need some additions to the GraphOptions code to allow for dependent options or linking several to be exclusive etc etc.