Fixed Width Dividers added
Option added in the same fashion as the fixed size dividers. This one puts the lines at the leftmost point on the graph and at each distance specified by the option from that point. Labels and lines are filled in using the same d3 code as the other options, except for the hanging runoff label that is between the last breakpoint and the end of the graph in the case that the intervals do not fit perfectly.