Beginning of Make Your Own Groups
Make Your Own Groups option added, when enable it makes a divider line at the beginning and end of the x-axis. A square tab is at the top of the start line, when clicked and dragged it creates a new divider line which follows the mouse until dropped.
STILL NEED TO:
 - Keep track of created divider lines so they dont disappear on graph redraw
- Calculate group populations based on lines that have been created and dropped
- Add drag handlers to lines created from the starting line so they can be adjusted after creatiion