Styled population labels a bit, bug fix
Population labels for groups now change between red and green if they are empty or not, respectively. Population labels are now bolded.
Fixed a bug in the fixed size groups option where if the total number of points was evenly divisible by the chosen group size, then an extra division population of 0 was added to the graph despite not having breakpoints to be positioned between.