Allow deletion of MYOG flags
Flags can now be dragged off the right side of a graph(at most 20px) at which point a message will appear informing the user that, if dropped, the flag will be deleted. Since id's of divLines are associated with their index in the customDivs array, instead of deleting entries from the array and having to change how customDivs is accessed, any deleted flags just have their entry set to -Infinity in customDivs. When population labels are drawn, any entries in customDivs with that value are ignored.