Fixed ClickToAddData
Data added to the assigned column now properly shows up in the spreadsheet and dataset. To accomplish this blank data (i.e. 0's) had to be added to the end of all other columns in the dataset that had the click added to it. This is fairly undesirable as those 0's will still register as data, looking for a way to add just null values to the other columns and still retain functionality.
If more than two datacolumns are assigned to a dot plot, then clicking will display an alert openTip saying that clicking to add data is disabled so as to prevent confusion over what column data would be assigned to etc.