ClickToAdd adds null entries to dataset, changed SS cell rendering
Instead of adding entries of '0' into the newly created row of clickToAdd, 'null' is added so that there isn't extra unexpected data showing up in graphs, since 'null
 entries are likely to be ignored by graphs.
Changed coloring of spreadsheet cells from relying on css classes based on the cell being numeric or not. Now the HandsOnTable instance uses a built-in HoT renderer function which colors cells.
This new renderer colors as follows:
   -Blank or null entries are colored a light gray
   -If the data column is numeric, cells are colored the same salmonish color as before.
   -Else the cell is colored the categorical green.