Fixed MYOG flag dragging and added scale customization
Dropping a Make Your Own Group flag no longer drops a new data point beneath the mouse. Needed to stop propagation of mousedown and click events from reaching the svg. Also added an invisible mask circle which follows the mouse so that if the mouse wasn't actually over the flag on drop, the click would be over some element within the g of the graph instead of over the empty space of the svg.
Added options for changing the graph's scale. Default values are drawn from the minMaxWPadding function already in the graph. Changed the initialization of the x scale in setViewMembers so that it uses the options values instead of the return from minMaxWPadding. Also changed the BoxPlot Drawer so that it drew its domain from the domain of the x scale instead of from the return of minMaxWPadding since the domain of the x scale will always be updated from the options.