Added option to show Least Squares when regression line is on
New option "Show Squares" is added to the options menu. Changed drawLeastSquares function to take another boolean argument for whether or not it should draw the squares along with the line. 
After the check of that boolean near the end of the function, it now goes through the data and creates a squares array.
Inside this array are attributes for the x and y coordinates of the anchor point on the regression line for each data point in the graph, the absolute length, in pixels, of the sides of the square, and a check of whether the reference point is above the regression line or below the regression line. The squares are then drawn using this information and some logic to determine how the rectangles should be anchored around the line based on its slope and the above and below boolean.