Working on getting ellipse to maintain coordinates thru container resize
Points are now stored based on their graph coordinates in the graph-wide variable ellipsePts. When the graph is redrawn it uses these points and draws new rectangles and an ellipse using its new xScale and yScale to determine pixel placements for them in the new graph container. So far this works fine for placements of everything, vertices and the center of the ellipse are showing up in the same spots, e.g. a vertex that was at the graph coordinate (2,2) will still be at (2,2) after a container resize. However, the calculation for determining the angle of rotation of the ellipse using new scales is a bit off when drawn in a new space after changing the angle of rotation from the default 0. However, when reverting back to the original size of when the angle change was made will rotate the ellipse correctly, even though it should be the same process to determine the rotation angle. Will look into this further.

An anchor rectangle in the center of the ellipse was added that can be used to drag the entire ellipse around, maintaining its shape and angle of rotation.