Updated ellipse, should work totally
Resizing doesn't play well with maintaining an ellipse, so any resizes to the graph's container result in the ellipse being reset to the default shape and position in the new container. 
Added the detection of points within the ellipse. Implemented using the equation described in the top answer of this post http://math.stackexchange.com/questions/426150/what-is-the-general-equation-of-the-ellipse-that-is-not-in-the-origin-and-rotate