Added median median line
Has all options from old seeIt minus the display of the equation for the line. These options are currently always on when the graph option Median-Median is on, so the divisions and median points are shown along with the line.
Moved some lines around so that when the graph is drawn the ellipse is the last thing drawn, ensuring that if other options that draw on top of the graph are enabled, the ellipse is still on top and can be dragged for resizing.