Added option for user movable Least Squares line
Option added is "Draw Your Least Squares Line".
Currently will draw the default initial line which has a slope of 0 and a y-intercept at the halfway point of the graphs y-axis.
Modified the minMax function from the last commit so that an option can be handed to it to have it return simply the minimum and maximum x and y values of the graph instead of the padded and rounded versions used by the axes.
Added the equation for the Median-Median line to show up in the top left corner of the graph when the Median-Median is enabled.