Made ends of the line draggable and made squares update correctly
Squares are now drawn only once and on drag the attributes of every square are changed. This was done so that the draggable points of the line are always on top of the squares so that they can be properly clicked. NOTE THAT the points of the plot that are beneath any least squares cannot be dragged, though this may not be a usage issue since the least squares line the user would be moving would be the major focus.
Switched up some style of the squares and lines to make them more consistent with the old SeeIt.