User Least Squares Line is can now be translated from the middle
Added in the draggable square in the middle of the line which will move the entire line around while maintaining its same slope.
Added the calculation of the Sum of Squares relative to the users line, appears next to the equation of the line in the top left of the graph container.