Fixed graphs that use options to determine scale
Grpahs that use options to determine the scale of their axes (correlation and distribution dot plot) now actually update the graph options if data is assigned which is outside of its previous scale. This is done through the event 'option:update', triggered from within the graph. It goes up to the GraphView, which passes the event to the GraphOptions, which then updates the value in the option field. The argument passed to the event is an array of updates, so multiple options can be changed at once. Once all updates are completed the GraphOptions triggers the 'graph:update' event which is the same event that is triggered as when the save button is clicked in the option menu. This redraws the graph with the correct, updated options.
This is now the expected behavior for when data is assigned or changed for the correlation plot, and when data is assigned, changed, or destroyed in the distribution plot. Destroyed data is not accounted for in the correlation because both of its data roles are singular, so if any data is destroyed the graph will not be drawn.