Added buttons as a type of graph option. Now used to start/stop rotation
New graph option type! Buttons will trigger an event in the graph that passes the label of the option and the values of other options(like other graph events). The graph module will handle all option buttons in one eventCallback, even if they are different labels, so the module must deal with checks and different behavior from different buttons if there are multiples.
Spirograph now uses the Start and Stop option buttons to begin rotations and stop the transitions. If stopped, start will not pick up from where it left off, rather it will restart the full orbit pattern from the beginning.
Orbit lines should not be left on the graph when hitting the start button or resizing/etc. while an orbit pattern is already in progress.