Graph Maximize behavior updated
The maximize button for a graph now puts the graph in a "presentation mode" where its view occupies essentially the whole screen. Header and footer are hidden as well as the navbar, spreadsheet, and data sections of SeeIt. Buttons are added to the top left and right of the graph view to open/close the graph options menu and to exit presentation mode, respectively.