Added more options to Correlation Plot
These options are all the fine control options for the other options of the Correlation Plot. Option lines like median-median or the least squares or draw your own least squares can now be drawn with or without the detail squares or points accompanying them. This should complete all the options that we wanted to carry over from the old SeeIt3 correlations plot.